fix(OfflineFire): Paycheck for fired employees
Describe Pull request
First, make sure you've read and are following the contribution guidelines and style guide and your code reflects that.
Write up a clear and concise description of what your pull request adds or fixes and if it's an added feature explain why you think it should be included in the core.

If this is not pulled everyone fired will be getting 500 dollars paychecks. I had people abusing unemployment because of it. In my community the paycheck is 35 for unemployment so 500 is a huge jump and breaking the economy. This pull request pulls the salary dynamically from the shared.lua so it can adapt for everyone community. 

If your PR is to fix an issue mention that issue here

Questions (please complete the following information):

Have you personally loaded this code into an updated qbcore project and checked all it's functionality? [yes/no] (Be honest)
-yes
Does your code fit the style guidelines? [yes/no]
yes
Does your PR fit the contribution guidelines? [yes/no]
yes